+++ base/src/main/java/proguard/KotlinMetadataAdapter.java
@@ -10,6 +10,7 @@ package proguard;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import proguard.classfile.io.kotlin.KotlinMetadataWriter;
+import proguard.classfile.kotlin.KotlinMetadataVersion;
 import proguard.classfile.kotlin.visitor.ReferencedKotlinMetadataVisitor;
 import proguard.classfile.visitor.ClassCounter;
 import proguard.pass.Pass;
@@ -20,12 +21,10 @@ implements Pass
 {
     private static final Logger logger = LogManager.getLogger(KotlinMetadataAdapter.class);
 
-    private final Configuration configuration;
-
-    public KotlinMetadataAdapter(Configuration configuration)
-    {
-        this.configuration = configuration;
-    }
+    // TODO(T14074): Fix the Kotlin metadata version until Kotlin 1.7 is released. Normally, ProGuardCORE
+    //               will write out the kotlinx.metadata "compatible" version, which for 0.3 is 1.5.1.
+    //               ProGuardCORE 8.0.7 uses kotlinx.metadata 0.4.1 so fix the written version until 1.7 is released.
+    public static final KotlinMetadataVersion KOTLIN_METADATA_VERSION = new KotlinMetadataVersion(1, 5, 1);
 
 
     @Override
@@ -36,7 +35,11 @@ implements Pass
         ClassCounter counter = new ClassCounter();
         appView.programClassPool.classesAccept(
             new ReferencedKotlinMetadataVisitor(
-            new KotlinMetadataWriter((clazz, message) -> logger.warn(message), counter)));
+            new KotlinMetadataWriter(
+                (clazz, message) -> logger.warn(clazz.getName(), message),
+                KOTLIN_METADATA_VERSION,
+                counter
+            )));
 
         logger.info("  Number of Kotlin classes adapted:              " + counter.getCount());
     }