+++ base/src/main/java/proguard/ProGuard.java
@@ -132,6 +132,8 @@ public class ProGuard
                 mark();
             }
 
+            checkConfigurationAfterInitialization();
+
             if (configuration.addConfigurationDebugging)
             {
                 // Remember the initial state of the program classpool and resource filepool
@@ -357,6 +359,13 @@ public class ProGuard
         passRunner.run(new KotlinAnnotationStripper(configuration), appView);
     }
 
+    /**
+     * Checks the configuration after it has been initialized.
+     */
+    private void checkConfigurationAfterInitialization() throws Exception
+    {
+        passRunner.run(new AfterInitConfigurationChecker(configuration), appView);
+    }
 
     /**
      * Replaces primitive array initialization code by primitive array constants.
@@ -503,7 +512,7 @@ public class ProGuard
      */
     private void adaptKotlinMetadata() throws Exception
     {
-        passRunner.run(new KotlinMetadataAdapter(configuration), appView);
+        passRunner.run(new KotlinMetadataAdapter(), appView);
     }