+++ base/src/main/java/proguard/util/kotlin/asserter/constraint/MultiFilePartIntegrity.java
@@ -35,7 +35,7 @@ extends      AbstractKotlinMetadataConstraint
     public void visitKotlinMultiFilePartMetadata(Clazz clazz,
                                                  KotlinMultiFilePartKindMetadata kotlinMultiFilePartKindMetadata)
     {
-        AssertUtil util = new AssertUtil("Multi-file part " + clazz.getName(), reporter);
+        AssertUtil util = new AssertUtil("Multi-file part " + clazz.getName(), reporter, programClassPool, libraryClassPool);
         util.reportIfNullReference("referenced facade class", kotlinMultiFilePartKindMetadata.referencedFacadeClass);
 /*        new AssertUtil("Multi-file part " + clazz.getName(), reporter).
             reportIfNullReference("referenced module", kotlinMultiFilePartKindMetadata.referencedModule);*/