+++ base/src/main/java/proguard/util/kotlin/asserter/constraint/PropertyIntegrity.java
@@ -46,7 +46,7 @@ extends      AbstractKotlinMetadataConstraint
                                  KotlinDeclarationContainerMetadata kotlinDeclarationContainerMetadata,
                                  KotlinPropertyMetadata kotlinPropertyMetadata)
     {
-        AssertUtil util = new AssertUtil("Property " + kotlinPropertyMetadata.name, reporter);
+        AssertUtil util = new AssertUtil("Property " + kotlinPropertyMetadata.name, reporter, programClassPool, libraryClassPool);
 
         util.reportIfNull("backingFieldSignature, getterSignature or setterSignature",
                           kotlinPropertyMetadata.backingFieldSignature,
@@ -56,6 +56,8 @@ extends      AbstractKotlinMetadataConstraint
         if (kotlinPropertyMetadata.backingFieldSignature  != null)
         {
             util.reportIfNullReference("backing field class",
+                                       kotlinPropertyMetadata.referencedBackingFieldClass);
+            util.reportIfClassDangling("backing field class",
                                        kotlinPropertyMetadata.referencedBackingFieldClass);
             util.reportIfNullReference("backing field",
                                        kotlinPropertyMetadata.referencedBackingField);
@@ -78,11 +80,27 @@ extends      AbstractKotlinMetadataConstraint
         {
             util.reportIfNullReference("synthetic annotations method class",
                                        kotlinPropertyMetadata.referencedSyntheticMethodClass);
+            util.reportIfClassDangling("synthetic annotations method class",
+                                       kotlinPropertyMetadata.referencedSyntheticMethodClass);
             util.reportIfNullReference("synthetic annotations method",
                                        kotlinPropertyMetadata.referencedSyntheticMethodForAnnotations);
             util.reportIfMethodDangling("synthetic annotations method",
                                         kotlinPropertyMetadata.referencedSyntheticMethodClass,
                                         kotlinPropertyMetadata.referencedSyntheticMethodForAnnotations);
+        }
+
+        if (kotlinPropertyMetadata.syntheticMethodForDelegate != null)
+        {
+            util.reportIfNullReference("synthetic delegate method class",
+                                       kotlinPropertyMetadata.referencedSyntheticMethodForDelegateClass);
+            util.reportIfClassDangling("synthetic delegate method class",
+                                       kotlinPropertyMetadata.referencedSyntheticMethodForDelegateClass);
+            util.reportIfNullReference("synthetic delegate method",
+                                       kotlinPropertyMetadata.referencedSyntheticMethodForDelegateMethod);
+            util.reportIfMethodDangling("synthetic delegate method",
+                                        kotlinPropertyMetadata.referencedSyntheticMethodForDelegateClass,
+                                        kotlinPropertyMetadata.referencedSyntheticMethodForDelegateMethod);
+
         }
     }
 }